Rework priority lookup, add namebased prio lookup
Rework priority lookup so that status data can be used for
looking up priorities, so that name based lookup method can be
implemented. Using this priorities can be detected based on the
status data, using dynamic regex based pattern/replacement matches
defined in the configuration file.
Changed the internal structure of the alert_history and priority
hashes to use the same notation.