Get rid of JS based resizing magic and replace with CSS
Get rid of JS based trickery for scaling the tile contents.
Now only the font size of the .tile-frame element is scaled relative
to the size of the tile (can have multiple columns), the rest is
done via CSS and 'em' units.
Slight cosmetics on the title and header to match the old size and
layout